message

+++ src/main/java/com/example/microserviceCourse/repository/UserRepository.java
@@ -1,12 +1,16 @@
 package com.example.microserviceCourse.repository;
 
 import com.example.microserviceCourse.user.User;
+import jakarta.persistence.metamodel.SingularAttribute;
+import org.springframework.data.jpa.domain.AbstractPersistable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.io.Serializable;
 import java.util.Optional;
 
 @Repository
 public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> getAllBy(int id);
+    Optional<User> getAllBy(Long id);
+
 }
